+++ scripts/aws_run.py
@@ -22,8 +22,8 @@ from weldx.asdf.tags.weldx.aws.process.shielding_gas_for_procedure import (
     ShieldingGasForProcedure,
 )
 from weldx.asdf.tags.weldx.aws.process.shielding_gas_type import ShieldingGasType
-from weldx.asdf.tags.weldx.core.iso_groove import get_groove
 from weldx.constants import WELDX_QUANTITY as Q_
+from weldx.welding.groove.iso_9692_1 import get_groove
 
 # welding process -----------------------------------------------------------------
 gas_comp = [
