+++ weldx/__init__.py
@@ -9,9 +9,6 @@ import weldx.geometry
 import weldx.transformations
 import weldx.utility
 import weldx.welding
-
-# class imports to weldx namespace
-from weldx.asdf.tags.weldx.core.iso_groove import get_groove
 from weldx.constants import WELDX_QUANTITY as Q_
 from weldx.core import MathematicalExpression, TimeSeries
 from weldx.geometry import ArcSegment, Geometry, LineSegment, Profile, Shape, Trace
@@ -20,6 +17,9 @@ from weldx.transformations import (
     LocalCoordinateSystem,
     WXRotation,
 )
+
+# class imports to weldx namespace
+from weldx.welding.groove.iso_9692_1 import get_groove
 
 # versioneer
 from ._version import get_versions
