+++ weldx/asdf/tags/weldx/core/__init__.py
@@ -2,7 +2,6 @@ from . import (
     common_types,
     data_array,
     dataset,
-    iso_groove,
     mathematical_expression,
     time_series,
     transformations,
